Fix Kotlin errors: callback params and Long type mismatch

+++ app/src/main/java/com/ffai/assistant/network/ServerConnection.kt
@@ -41,8 +41,8 @@ class ServerConnection {
             })
         }
         install(HttpTimeout) {
-            connectTimeoutMillis = ServerConfig.CONNECTION_TIMEOUT.toInt()
-            requestTimeoutMillis = ServerConfig.RESPONSE_TIMEOUT.toInt()
+            connectTimeoutMillis = ServerConfig.CONNECTION_TIMEOUT
+            requestTimeoutMillis = ServerConfig.RESPONSE_TIMEOUT
         }
     }
